Split rest methods in mulitple files
