+++ src/Weaviate.Client/Rest/Client.cs
@@ -6,86 +6,7 @@ using Weaviate.Client.Rest.Dto;
 
 namespace Weaviate.Client.Rest;
 
-public static class HttpResponseMessageExtensions
-{
-    public static async Task EnsureExpectedStatusCodeAsync(
-        this HttpResponseMessage response,
-        SortedSet<HttpStatusCode> codes,
-        string error = ""
-    )
-    {
-        if (codes.Contains(response.StatusCode))
-        {
-            return;
-        }
-
-        var content = await response.Content.ReadAsStringAsync();
-
-        if (response.Content != null)
-            response.Content.Dispose();
-
-        content =
-            $"Unexpected status code {response.StatusCode}. Expected: {string.Join(", ", codes)}. {error}. Server replied: {content}";
-
-        throw new SimpleHttpResponseException(response.StatusCode, codes, content);
-    }
-
-    public static async Task EnsureExpectedStatusCodeAsync(
-        this HttpResponseMessage response,
-        SortedSet<int> codes,
-        string error = ""
-    )
-    {
-        await EnsureExpectedStatusCodeAsync(
-            response,
-            [.. codes.Select(x => (HttpStatusCode)x)],
-            error
-        );
-    }
-
-    public static async Task EnsureExpectedStatusCodeAsync(
-        this HttpResponseMessage response,
-        int code,
-        string error = ""
-    )
-    {
-        await EnsureExpectedStatusCodeAsync(response, [(HttpStatusCode)code], error);
-    }
-
-    public static async Task EnsureExpectedStatusCodeAsync(
-        this HttpResponseMessage response,
-        HttpStatusCode code,
-        string error = ""
-    )
-    {
-        await EnsureExpectedStatusCodeAsync(response, [code], error);
-    }
-
-    public static async Task EnsureSuccessStatusCodeAsync(this HttpResponseMessage response)
-    {
-        await EnsureExpectedStatusCodeAsync(response, [HttpStatusCode.OK]);
-    }
-}
-
-internal class SimpleHttpResponseException : WeaviateServerException
-{
-    public HttpStatusCode StatusCode { get; private set; }
-    public ISet<HttpStatusCode> ExpectedStatusCodes { get; private set; } =
-        new SortedSet<HttpStatusCode>();
-
-    public SimpleHttpResponseException(
-        HttpStatusCode statusCode,
-        SortedSet<HttpStatusCode> expectedStatusCodes,
-        string content
-    )
-        : base(content)
-    {
-        StatusCode = statusCode;
-        ExpectedStatusCodes = expectedStatusCodes;
-    }
-}
-
-public class WeaviateRestClient : IDisposable
+public partial class WeaviateRestClient : IDisposable
 {
     private readonly bool _ownershipClient;
 
@@ -122,352 +43,5 @@ public class WeaviateRestClient : IDisposable
         {
             _httpClient.Dispose();
         }
-    }
-
-    internal async Task<Dto.Schema?> CollectionList()
-    {
-        var response = await _httpClient.GetAsync(WeaviateEndpoints.Collection());
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection list");
-
-        var contents = await response.Content.ReadFromJsonAsync<Dto.Schema>(
-            options: RestJsonSerializerOptions
-        );
-
-        return contents;
-    }
-
-    internal async Task<Dto.Class?> CollectionGet(string name)
-    {
-        var response = await _httpClient.GetAsync(WeaviateEndpoints.Collection(name));
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection get");
-
-        if (response.Content.Headers.ContentLength == 0)
-        {
-            return default;
-        }
-
-        var contents = await response.Content.ReadFromJsonAsync<Dto.Class>(
-            options: RestJsonSerializerOptions
-        );
-
-        if (contents is null)
-        {
-            throw new WeaviateRestException();
-        }
-
-        return contents;
-    }
-
-    internal async Task CollectionDelete(string name)
-    {
-        var response = await _httpClient.DeleteAsync(WeaviateEndpoints.Collection(name));
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection delete");
-    }
-
-    internal async Task<Dto.Class> CollectionCreate(Dto.Class collection)
-    {
-        var response = await _httpClient.PostAsJsonAsync(
-            WeaviateEndpoints.Collection(),
-            collection,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection create");
-
-        var contents = await response.Content.ReadFromJsonAsync<Dto.Class>(
-            options: RestJsonSerializerOptions
-        );
-
-        if (contents is null)
-        {
-            throw new WeaviateRestException();
-        }
-
-        return contents;
-    }
-
-    internal async Task<Dto.Class> CollectionUpdate(string collectionName, Dto.Class collection)
-    {
-        var response = await _httpClient.PutAsJsonAsync(
-            WeaviateEndpoints.Collection(collectionName),
-            collection,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection update");
-
-        var contents =
-            await response.Content.ReadFromJsonAsync<Dto.Class>(options: RestJsonSerializerOptions)
-            ?? throw new WeaviateRestException();
-
-        return contents;
-    }
-
-    internal async Task<Dto.Object> ObjectInsert(Dto.Object data)
-    {
-        var response = await _httpClient.PostAsJsonAsync(
-            WeaviateEndpoints.Objects(),
-            data,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "insert object");
-
-        return await response.Content.ReadFromJsonAsync<Dto.Object>()
-            ?? throw new WeaviateRestException();
-    }
-
-    internal async Task<Dto.Object> ObjectReplace(string collectionName, Dto.Object data)
-    {
-        ArgumentNullException.ThrowIfNull(data.Id, nameof(data.Id));
-
-        var response = await _httpClient.PutAsJsonAsync(
-            WeaviateEndpoints.CollectionObject(collectionName, data.Id!.Value),
-            data,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "replace object");
-
-        return await response.Content.ReadFromJsonAsync<Dto.Object>()
-            ?? throw new WeaviateRestException();
-    }
-
-    internal async Task DeleteObject(string collectionName, Guid id, string? tenant = null)
-    {
-        var url = WeaviateEndpoints.CollectionObject(collectionName, id, tenant);
-        var response = await _httpClient.DeleteAsync(url);
-
-        await response.EnsureExpectedStatusCodeAsync([204, 404], "delete object");
-    }
-
-    internal async Task ReferenceAdd(
-        string collectionName,
-        Guid from,
-        string fromProperty,
-        Guid to,
-        string? tenant = null
-    )
-    {
-        var path = WeaviateEndpoints.Reference(collectionName, from, fromProperty, tenant);
-
-        var beacons = ObjectHelper.MakeBeacons(to);
-        var reference = beacons.First();
-
-        var response = await _httpClient.PostAsJsonAsync(
-            path,
-            reference,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "reference add");
-    }
-
-    internal async Task ReferenceReplace(
-        string collectionName,
-        Guid from,
-        string fromProperty,
-        Guid[] to,
-        string? tenant = null
-    )
-    {
-        var path = WeaviateEndpoints.Reference(collectionName, from, fromProperty, tenant);
-
-        var beacons = ObjectHelper.MakeBeacons(to);
-        var reference = beacons;
-
-        var response = await _httpClient.PutAsJsonAsync(
-            path,
-            reference,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "reference replace");
-    }
-
-    internal async Task ReferenceDelete(
-        string collectionName,
-        Guid from,
-        string fromProperty,
-        Guid to,
-        string? tenant = null
-    )
-    {
-        var path = WeaviateEndpoints.Reference(collectionName, from, fromProperty, tenant);
-
-        var beacons = ObjectHelper.MakeBeacons(to);
-        var reference = beacons.First();
-
-        var request = new HttpRequestMessage(HttpMethod.Delete, path);
-        request.Content = JsonContent.Create(
-            reference,
-            mediaType: null,
-            options: RestJsonSerializerOptions
-        );
-
-        var response = await _httpClient.SendAsync(request);
-
-        await response.EnsureExpectedStatusCodeAsync([200], "reference delete");
-    }
-
-    internal async Task CollectionAddProperty(string collectionName, Property property)
-    {
-        var path = WeaviateEndpoints.CollectionProperties(collectionName);
-
-        var response = await _httpClient.PostAsJsonAsync(
-            path,
-            property,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection property add");
-    }
-
-    internal async Task<BatchReferenceResponse[]> ReferenceAddMany(
-        string collectionName,
-        Models.DataReference[] references,
-        string? tenant = null,
-        ConsistencyLevels? consistencyLevel = null
-    )
-    {
-        var batchRefs = references.SelectMany(r =>
-            ObjectHelper
-                .MakeBeacons(r.To)
-                .SelectMany(b => b.Values)
-                .Select(beacon => new BatchReference
-                {
-                    From = new Uri(
-                        ObjectHelper.MakeBeaconSource(collectionName, r.From, r.FromProperty)
-                    ),
-                    To = new Uri(beacon),
-                    Tenant = tenant ?? default!,
-                })
-        );
-
-        var path = WeaviateEndpoints.ReferencesAdd(consistencyLevel);
-
-        var response = await _httpClient.PostAsJsonAsync(
-            path,
-            batchRefs,
-            options: RestJsonSerializerOptions
-        );
-
-        await response.EnsureExpectedStatusCodeAsync([200], "reference add many");
-
-        return await response.Content.ReadFromJsonAsync<BatchReferenceResponse[]>(
-                RestJsonSerializerOptions
-            ) ?? throw new WeaviateRestException();
-    }
-
-    internal async Task<bool> CollectionExists(object collectionName)
-    {
-        var path = WeaviateEndpoints.Collection();
-
-        var response = await _httpClient.GetAsync(path);
-
-        await response.EnsureExpectedStatusCodeAsync([200], "collection property add");
-
-        var schema = await response.Content.ReadFromJsonAsync<Schema>(
-            options: RestJsonSerializerOptions
-        );
-
-        return schema?.Classes?.Any(c => c.Class1 is not null && c.Class1!.Equals(collectionName))
-            ?? false;
-    }
-
-    internal async Task<Dto.Meta?> GetMeta()
-    {
-        var path = WeaviateEndpoints.Meta();
-
-        var response = await _httpClient.GetAsync(path);
-
-        await response.EnsureExpectedStatusCodeAsync([200], "get meta endpoint");
-
-        var meta = await response.Content.ReadFromJsonAsync<Meta>(
-            options: RestJsonSerializerOptions
-        );
-
-        return meta;
-    }
-
-    internal async Task<NodeStatus[]> Nodes(string? collection, string verbosity)
-    {
-        var path = WeaviateEndpoints.Nodes(collection, verbosity);
-        var response = await _httpClient.GetAsync(path);
-        var responseContent = await response.Content.ReadAsStringAsync();
-        // Print the response content
-        Console.WriteLine(responseContent);
-        await response.EnsureExpectedStatusCodeAsync([200], "get nodes");
-        var nodes = await response.Content.ReadFromJsonAsync<NodesStatusResponse>(
-            options: RestJsonSerializerOptions
-        );
-        return nodes?.Nodes?.ToArray() ?? Array.Empty<NodeStatus>();
-    }
-
-    // Tenants API
-    internal async Task<IEnumerable<Rest.Dto.Tenant>> TenantsAdd(
-        string collectionName,
-        params Rest.Dto.Tenant[] tenants
-    )
-    {
-        var path = WeaviateEndpoints.CollectionTenants(collectionName);
-
-        var response = await _httpClient.PostAsJsonAsync(
-            path,
-            tenants,
-            options: RestJsonSerializerOptions
-        );
-        await response.EnsureExpectedStatusCodeAsync([200], "tenants add");
-        var result = await response.Content.ReadFromJsonAsync<IEnumerable<Rest.Dto.Tenant>>(
-            options: RestJsonSerializerOptions
-        );
-        return result ?? Enumerable.Empty<Rest.Dto.Tenant>();
-    }
-
-    internal async Task<IEnumerable<Rest.Dto.Tenant>> TenantUpdate(
-        string collectionName,
-        params Rest.Dto.Tenant[] tenants
-    )
-    {
-        if (tenants.Any(t => t.Name is null))
-        {
-            throw new ArgumentNullException(nameof(tenants), "Tenant names cannot be null.");
-        }
-
-        var path = WeaviateEndpoints.CollectionTenants(collectionName);
-        var response = await _httpClient.PutAsJsonAsync(
-            path,
-            tenants,
-            options: RestJsonSerializerOptions
-        );
-        await response.EnsureExpectedStatusCodeAsync([200], "tenant update");
-
-        var result = await response.Content.ReadFromJsonAsync<Rest.Dto.Tenant[]>(
-            options: RestJsonSerializerOptions
-        );
-
-        return result ?? Enumerable.Empty<Rest.Dto.Tenant>();
-    }
-
-    internal async Task TenantsDelete(string collectionName, IEnumerable<string> tenantNames)
-    {
-        if (tenantNames.Any(name => string.IsNullOrWhiteSpace(name)))
-        {
-            throw new ArgumentException(
-                "Tenant names cannot be null or empty.",
-                nameof(tenantNames)
-            );
-        }
-
-        var path = WeaviateEndpoints.CollectionTenants(collectionName);
-        var request = new HttpRequestMessage(HttpMethod.Delete, path)
-        {
-            Content = JsonContent.Create(tenantNames, options: RestJsonSerializerOptions),
-        };
-        var response = await _httpClient.SendAsync(request);
-        await response.EnsureExpectedStatusCodeAsync([200], "tenants delete");
     }
 }
